+++ src/unistr.cpp
@@ -145,13 +145,14 @@ void init_unistr(py::module &m, py::class_<Replaceable, UObject> &rep, py::class
   us.def("__hash__", &UnicodeString::hashCode);
 
   us.def(
-      "__iadd__",
-      [](UnicodeString &self, const _UnicodeStringVariant &other) { return self += VARIANT_TO_UNISTR(other); },
-      py::is_operator(), py::arg("other"));
-
-  us.def(
-      "__iadd__", [](UnicodeString &self, UChar32 other) { return self += other; }, py::is_operator(),
-      py::arg("other"));
+        "__iadd__",
+        [](UnicodeString &self, const _UnicodeStringVariant &other) -> UnicodeString & {
+          return self += VARIANT_TO_UNISTR(other);
+        },
+        py::is_operator(), py::arg("other"))
+      .def(
+          "__iadd__", [](UnicodeString &self, UChar32 other) -> UnicodeString & { return self += other; },
+          py::is_operator(), py::arg("other"));
 
   us.def(
       "__le__",